// micro-services/core/src/enums/entityTypeV2.enum.ts
export enum EntityTypeV2 {
  User = "entity_user",
  Organization = "entity_organization",
  OrganizationMembership = "entity_organization_membership",
  Team = "entity_team",
  TeamMembership = "entity_team_membership",
  Group = "entity_group",
  GroupMembership = "entity_group_membership",
  Meeting = "entity_meeting",
  MeetingMembership = "entity_meeting_membership",
  OneOnOneMembership = "entity_one_on_one_membership",
  PendingMessage = "entity_pending_message",
  Message = "enttiy_message",
  PendingInvitation = "entity_pending_invitation",
}